fix issue #94: bad compilation of cut inside if-then port to arm64/Windows using clang and MSVC (see src/WINDOWS) improve all Windows/msys ports improve Windows Inno setup installer improve temporary_name/2 and temporary_file/2 randomness (mitigate race conditions) fix issue #93: write_term_to_chars/3 and friends incorrectly instantiate variables fix issue #71: phrase/2 too eager execution fix bug in riscv64 mapper - now gprolog can be compiled with clang rename hexfilter -> gplc-filter (backward incompatibility) remove deprecated gplc option --no-decode-hexa (use --no-demangling) extend gplc to accept .S (preprocessed assembly files) gplc accepts gcc-like options -Wl, -Wa, -Wc, rewrite the C to native Prolog code trampoline in assembly engine1.c is replaced by trampoline.S (old engine1.c -> trampoline-old.c) include setjmp/longjmp replacement for Windows 64 bit in trampoline.S modify configure.in to preserve user-supplied CFLAGS on command-line add gprolog top-level shorthands: -q for --quiet and -g for --init-goal strengthen C code strictness (enabling extra compiler warnings) refactor unit initialization (a unit is a Prolog program compiled to native code) new unit_ctor.[ch] (replaces obj_chain) now based on a portable constructor facility (see ctor.h) modify randomize/0 to produce more variation across runs change byte-code emulator (BCI register replaced by extendend local environments, see bc_supp.c) fix issue #85: bug in the GC-clause (retracting current clause) fix and improve the Prolog debugger fix bug in pl2wam with multifile aux predicate (prevent int_overflow) extend format/2-3 to conform to Prolog PIP-0110 https://prolog-lang.org/ImplementersForum/0110-format.html add tab-padding support (~| ~+ ~t), remove handling of ~? fix issue #90: crash due to a bad handling of WAM call_c in wam2ma

+++ src/Tools/map_test.c
@@ -6,7 +6,7 @@
  * Descr.: test for the map data structure with red-black trees            *
  * Author: Daniel Diaz                                                     *
  *                                                                         *
- * Copyright (C) 1999-2025 Daniel Diaz                                     *
+ * Copyright (C) 1999-2026 Daniel Diaz                                     *
  *                                                                         *
  * This file is part of GNU Prolog                                         *
  *                                                                         *
@@ -402,11 +402,11 @@ void
 test2(void)
 {
   struct map_rbt map;
-  long seed = (long) time(NULL);
+  unsigned int seed = (unsigned int) time(NULL);
 
   //printf("\nTest 2\n");
   //printf("N: %d\n", N);
-  //printf("seed: %ld\n", seed);
+  //printf("seed: %u\n", seed);
   srand(seed);
